+++ approvalcrest-core/src/main/java/com/github/karsaig/approvalcrest/FileMatcherConfig.java
@@ -1,16 +1,22 @@
 package com.github.karsaig.approvalcrest;
 
-import static com.github.karsaig.approvalcrest.EnvVarReader.getBooleanProperty;
+import static com.github.karsaig.approvalcrest.EnvVarReader.getBooleanProperties;
 
 public class FileMatcherConfig {
 
     private static final String UPDATE_IN_PLACE_OLD_NAME = "jsonMatcherUpdateInPlace";
     private static final String UPDATE_IN_PLACE_NAME = "fileMatcherUpdateInPlace";
+    private static final String UPDATE_IN_PLACE_ALIAS = "fMUInPlace";
     private static final String PASS_ON_CREATE = "fileMatcherPassOnCreate";
+    private static final String PASS_ON_CREATE_ALIAS = "fMPOnCreate";
     private static final String BUILD_INDEX_NAME = "buildFileIndex";
+    private static final String BUILD_INDEX_ALIAS = "bFIndex";
     private static final String APPROVED_DIRECTORY_NAME = "useApprovedDirectory";
+    private static final String APPROVED_DIRECTORY_ALIAS = "uADirectory";
     private static final String SORT_INPUT_FILE = "sortInputFile";
+    private static final String SORT_INPUT_FILE_ALIAS = "sIFile";
     private static final String STRICT_FILE_MATCHING = "fileMatcherStrictFileMatching";
+    private static final String STRICT_FILE_MATCHING_ALIAS = "fMStrictMatching";
 
 
     private final boolean overwriteInPlaceEnabled;
@@ -21,12 +27,12 @@ public class FileMatcherConfig {
     private final boolean strictFileMatching;
 
     public FileMatcherConfig() {
-        overwriteInPlaceEnabled = getBooleanProperty(UPDATE_IN_PLACE_OLD_NAME) || getBooleanProperty(UPDATE_IN_PLACE_NAME);
-        passOnCreateEnabled = getBooleanProperty(PASS_ON_CREATE);
-        buildIndex = getBooleanProperty(BUILD_INDEX_NAME);
-        approvedDirectory = getBooleanProperty(APPROVED_DIRECTORY_NAME);
-        sortInputFile = getBooleanProperty(SORT_INPUT_FILE);
-        strictFileMatching = getBooleanProperty(STRICT_FILE_MATCHING, "true");
+        overwriteInPlaceEnabled = getBooleanProperties(null, UPDATE_IN_PLACE_OLD_NAME, UPDATE_IN_PLACE_NAME, UPDATE_IN_PLACE_ALIAS);
+        passOnCreateEnabled = getBooleanProperties(null, PASS_ON_CREATE, PASS_ON_CREATE_ALIAS);
+        buildIndex = getBooleanProperties(null, BUILD_INDEX_NAME, BUILD_INDEX_ALIAS);
+        approvedDirectory = getBooleanProperties(null, APPROVED_DIRECTORY_NAME, APPROVED_DIRECTORY_ALIAS);
+        sortInputFile = getBooleanProperties(null, SORT_INPUT_FILE, SORT_INPUT_FILE_ALIAS);
+        strictFileMatching = getBooleanProperties("true", STRICT_FILE_MATCHING, STRICT_FILE_MATCHING_ALIAS);
     }
 
     public FileMatcherConfig(boolean overwriteInPlaceEnabled, boolean passOnCreateEnabled, boolean buildIndex, boolean approvedDirectory, boolean sortInputFile, boolean strictMatching) {
@@ -46,6 +52,14 @@ public class FileMatcherConfig {
 
     public boolean isPassOnCreateEnabled() {
         return passOnCreateEnabled;
+    }
+
+    public boolean isBuildIndex() {
+        return buildIndex;
+    }
+
+    public boolean isApprovedDirectory() {
+        return approvedDirectory;
     }
 
     public boolean isSortInputFile() {